% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_automated.r
\name{ARIMAanalyse}
\alias{ARIMAanalyse}
\title{ARIMA model choice automation}
\usage{
ARIMAanalyse(seriex, laglim = 1, crossvalid = c(2015, 4), limit = 3,
  method = "ML")
}
\arguments{
\item{seriex}{A univariate time serie to be studied}

\item{laglim}{The maximum difference limit for I part (it is the same for ordinary and seasonal components)}

\item{crossvalid}{The point for cross-validation}

\item{limit}{A maximum number of models to be selected by criteria (3 in sample and 3 out of sample criterias are used)}

\item{method}{The estimation method to be applied (as in the timeserie package)}
}
\value{
A list of selected models estimated with SARIMA with additional performance information
\describe{
\item{crit_study}{A data.frame of main in-sample choice criterias for selected models}
\item{crit_test}{A data.frame of main out of sample choice criterias for selected models}
\item{plot_study}{Autoplot predictions for best in-sample models}
\item{plot_test}{Autoplot predictions for best out of sample models}
\item{t_stat}{Test statistics for tests}
\item{t_pval}{P-values for tests}
\item{models}{The results for selected models}
\item{plots}{Grid-arranged Acf and Pacf plots for each difference sequence}
}
}
\description{
This function allows to select by informational criteria and performance results one or a serie of different models allowing for best performance in and out of sample. The function is a wrapper for forecast and tseries packages.
}
\examples{
ARIMAanalyse(seriex, 
     laglim = 5, crossvalid = c(2015, 4), limit = 3, 
     method = "ML")
}
\keyword{MCOanalyse}
