% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_selection.r
\name{LASSOanalyse}
\alias{LASSOanalyse}
\title{A LASSO model selection function}
\usage{
LASSOanalyse(datax, dependent, laglim = 5, study.t0 = TRUE,
  folds = 10, modlim = 1, varlim = 10)
}
\arguments{
\item{datax}{A data.frame of explicative variables to be studied of length equal to explicative variable vector}

\item{dependent}{A dependent variable vector}

\item{laglim}{The model allows to automatically convert the dataset in order to study lagged dependecies}

\item{study.t0}{A logical value indicating whether the t0 period should be taken into account}

\item{folds}{Number of folds for crossvalidation}

\item{modlim}{A maximum number of additional models to be presented in output}

\item{varlim}{A maximal number of variables to figure in supplementary models}
}
\value{
A list of selected models estimated with OLS
\describe{
 \item{optimal}{The model for best lambda value, selected by cross validation}
 \item{other}{A list of other models in order of goodness of fit with limited regressors} 
}
}
\description{
This function allows to select by crossvalidation one or a serie of different models allowing for best performance. The function is a wrapper for glmnet package.
}
\examples{
LASSOanalyse(data, dependent, 
     laglim = 5, study.t0 = TRUE, folds = 10, 
     modlim = 1, varlim = 10)
}
\keyword{LASSOanalyse}
