% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_df.r
\name{test.df}
\alias{test.df}
\title{Seasonal data.frames exploration function}
\usage{
test.df(datax, season)
}
\arguments{
\item{datax}{A data.frame of seasonal data to be studied of length equal to explicative variable vector. The first column is advised to be year and the second to be "M" for monthly data, or "T" for trimestrial data.}

\item{season}{An indicator of seasonality type to be used as present in the data.frame.}
}
\value{
A data.frame with test results by variable.
}
\description{
This function allows to test stationarity of seasonal series present in data.frame as well as explore seasonal effects.
}
\examples{
test.df = function(datax, season)

}
\keyword{test.df}
