% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_automated.r
\name{ARIMAanalyse}
\alias{ARIMAanalyse}
\title{ARIMA model choice automation}
\usage{
ARIMAanalyse(seriex, orderlim = c(1, 1, 1), seasonallim = c(1, 1, 1),
  crossvalid = c(2015, 4), limit = 3, method = "ML",
  adf.plim = 0.1, partlim = 0.3)
}
\arguments{
\item{seriex}{A univariate time serie to be studied}

\item{seasonallim}{The maximum limit for P, D and Q for SARIMA part}

\item{crossvalid}{The point for cross-validation}

\item{limit}{A maximum number of models to be selected by criteria (3 in sample and 3 out of sample criterias are used)}

\item{method}{The estimation method to be applied (as in the timeserie package)}

\item{adf.plim}{ADF selection probability limit}

\item{partlim}{Part of autocorrelated lags in lags}

\item{order}{The maximum limit for p, d and q for ordinary ARIMA part}
}
\value{
A list of selected models estimated with SARIMA with additional performance information
}
\description{
This function allows to select by informational criteria and performance results one or a serie of different models allowing for best performance in and out of sample. The function is a wrapper for forecast and tseries packages.
}
\examples{
ARIMAanalyse = function(seriex, 
   orderlim = c(1, 1, 1), seasonallim = c(1, 1, 1), 
   crossvalid = c(2015, 4), limit = 3, 
   method = "ML", adf.plim = 0.1, partlim = 0.3)
   
}
\keyword{ARIMAanalyse}
